[refactor]: move computation backend to `scores.fast.fss`
