+++ src/scores/fast/fss/typing.py
@@ -1,0 +1,31 @@
+"""
+Specific type definitions for Fss backend    
+"""
+from enum import Enum
+from typing import TYPE_CHECKING, Union
+
+import numpy as np
+import numpy.typing as npt
+
+# Note: soft keyword `type` only support from >=3.10
+# "tuple" of 64 bit floats
+f8x3 = np.dtype("f8, f8, f8")
+# Note: `TypeAlias` on variables not available for python <3.10
+if TYPE_CHECKING:  # pragma: no cover
+    FssDecomposed = Union[np.ArrayLike, np.DtypeLike]
+else:  # pragma: no cover
+    FssDecomposed = npt.NDArray[f8x3]
+
+
+class FssComputeMethod(Enum):
+    """
+    Choice of compute backend for FSS.
+    - Standard
+        - NUMPY (default)
+    - Optimized
+        - NUMBA (currently unavailable)
+    """
+
+    INVALID = -1
+    NUMPY = 1  # (default)
+    NUMBA = 2
